@@ -1,24 +1,11 @@
 package ilugbom.org.in.batchmaker;
 
-import android.support.v7.app.AlertDialog;
 import android.content.DialogInterface;
 import android.os.Environment;
-import android.support.v4.content.FileProvider;
-import android.support.v7.app.AlertDialog;
-import android.text.InputType;
-import android.view.View;
-import android.view.inputmethod.InputMethodManager;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
+import androidx.core.content.FileProvider;
+import androidx.appcompat.app.AlertDialog;
 
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
 import java.util.ArrayList;
 import java.util.List;
 
